C++ Programming project
for my programming class. Beginning description.

You have been hired as a programmer for Blue Star Electric. The first program you must write generates a form letter to a customer with an overdue account. The letter will appear in the form below with the fields in brackets <> replaced by actual values.